Extract search feature from BrowserTree
Search is now implemented by MediaSearchEngine.
Media search is unrelated to the structure of the browser tree,
therefore it could be extracted to make BrowserTree simpler.
