+++ service/src/main/java/fr/nihilus/music/service/MusicService.kt
@@ -45,7 +45,7 @@ import fr.nihilus.music.core.playback.RepeatMode
 import fr.nihilus.music.core.settings.Settings
 import fr.nihilus.music.media.usage.UsageManager
 import fr.nihilus.music.service.browser.BrowserTree
-import fr.nihilus.music.service.browser.PaginationOptions
+import fr.nihilus.music.service.browser.MediaSearchEngine
 import fr.nihilus.music.service.browser.SearchQuery
 import fr.nihilus.music.service.notification.MediaNotificationBuilder
 import fr.nihilus.music.service.notification.NOW_PLAYING_NOTIFICATION
@@ -68,6 +68,7 @@ class MusicService : BaseBrowserService() {
 
     @Inject internal lateinit var dispatchers: AppDispatchers
     @Inject internal lateinit var browserTree: BrowserTree
+    @Inject internal lateinit var searchEngine: MediaSearchEngine
     @Inject internal lateinit var subscriptions: SubscriptionManager
     @Inject internal lateinit var notificationBuilder: MediaNotificationBuilder
     @Inject internal lateinit var usageManager: UsageManager
@@ -271,7 +272,7 @@ class MusicService : BaseBrowserService() {
         serviceScope.launch(dispatchers.Default) {
             val parsedQuery = SearchQuery.from(query, extras)
 
-            val searchResults = browserTree.search(parsedQuery)
+            val searchResults = searchEngine.search(parsedQuery)
             val builder = MediaDescriptionCompat.Builder()
             result.sendResult(searchResults.map { it.toItem(builder) })
         }
