+++ service/src/test/java/fr/nihilus/music/service/browser/TestDoubles.kt
@@ -115,13 +115,3 @@ internal object StubSpotifyManager : SpotifyManager {
 
     override fun sync(): Flow<SyncProgress> = stub()
 }
-
-internal object StubPlaylistDao : PlaylistDao() {
-    override val playlists: Flow<List<Playlist>> get() = stub()
-    override suspend fun findPlaylist(playlistId: Long): Playlist? = stub()
-    override fun getPlaylistTracks(playlistId: Long): Flow<List<PlaylistTrack>> = stub()
-    override suspend fun savePlaylist(playlist: Playlist): Long = stub()
-    override suspend fun addTracks(tracks: List<PlaylistTrack>): Unit = stub()
-    override suspend fun deletePlaylist(playlistId: Long): Unit = stub()
-    override suspend fun deletePlaylistTracks(trackIds: LongArray): Unit = stub()
-}
